+++ package.json
@@ -32,7 +32,6 @@
     "react-collapse": "^4.0.3",
     "react-dom": "^16.8.6",
     "react-js-pagination": "^3.0.2",
-    "react-markdown": "^4.2.2",
     "react-media": "^1.9.2",
     "react-motion": "^0.5.2",
     "react-redux": "^5.0.6",
