+++ src/project/Project.test.js
@@ -30,7 +30,7 @@ import { createMemoryHistory } from 'history';
 
 import { StateKind, StateModel } from '../model/Model';
 import Project from './Project';
-import { filterPaths, sanitizedHTMLFromMarkdown } from './Project.present'
+import { filterPaths } from './Project.present'
 import State, { ProjectModel } from  './Project.state';
 import { testClient as client } from '../api-client';
 import { slugFromTitle } from '../utils/HelperFunctions'
@@ -135,114 +135,5 @@ describe('path filtering', () => {
     expect(paths).toEqual(['foo.txt', 'bar', 'myfolder/.hidden', 'myfolder/visible',
     'myfolder/.alsohidden/other.txt',
     'myfolder/alsovisible/.hidden', 'myfolder/alsovisible/other.txt']);
-  })
-});
-
-describe('html sanitization', () => {
-  it('handles empty markdown', () => {
-    const markdown = ""
-    const html = sanitizedHTMLFromMarkdown(markdown)
-    expect(html).toEqual("")
-  })
-
-  it('handles pure markdown', () => {
-    const markdown = "# internal-test\nA Renku project.\n\nThis is an *internal* project that is used for testing."
-    const html = sanitizedHTMLFromMarkdown(markdown)
-    expect(html).toEqual(`<h1 id="internaltest">internal-test</h1>\n<p>A Renku project.</p>\n<p>This is an <em>internal</em> project that is used for testing.</p>`)
-  })
-
-  it('handles mixed markdown', () => {
-    const markdown =`# internal-test
-
-A Renku project.
-
-This is an *internal* project that is used for testing.
-
-<div>
-  This is some HTML in a div, since that is valid Markdown.
-</div>
-
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-sm-8">
-      An 8 unit column here on the left.
-    </div>
-    <div class="col-sm-4">
-      A 4 unit column here on the right.
-    </div>
-  </div>
-</div>`
-    const expected = `<h1 id="internaltest">internal-test</h1>
-<p>A Renku project.</p>
-<p>This is an <em>internal</em> project that is used for testing.</p>
-<div>
-  This is some HTML in a div, since that is valid Markdown.
-</div>
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-sm-8">
-      An 8 unit column here on the left.
-    </div>
-    <div class="col-sm-4">
-      A 4 unit column here on the right.
-    </div>
-  </div>
-</div>`
-    const html = sanitizedHTMLFromMarkdown(markdown)
-    expect(html).toEqual(expected)
-  })
-
-  it('strips suspicious markdown', () => {
-    const markdown =`# internal-test
-
-A Renku project.
-
-This is an *internal* project that is used for testing.
-
-<div>
-  This is some HTML in a div, since that is valid Markdown.
-</div>
-
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-sm-8">
-      An 8 unit column here on the left.
-    </div>
-    <div class="col-sm-4">
-      A 4 unit column here on the right.
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-sm-8">
-      The following alerts should get sanitized away.
-      <script>alert('xss');</script>
-      hello <a name="n" href="javascript:alert('xss')">*you*</a>
-    </div>
-  </div>
-</div>`
-    const expected = `<h1 id="internaltest">internal-test</h1>
-<p>A Renku project.</p>
-<p>This is an <em>internal</em> project that is used for testing.</p>
-<div>
-  This is some HTML in a div, since that is valid Markdown.
-</div>
-<div class="container-fluid">
-  <div class="row">
-    <div class="col-sm-8">
-      An 8 unit column here on the left.
-    </div>
-    <div class="col-sm-4">
-      A 4 unit column here on the right.
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-sm-8">
-      The following alerts should get sanitized away.
-      \n      hello <a name="n">*you*</a>
-    </div>
-  </div>
-</div>`
-    const html = sanitizedHTMLFromMarkdown(markdown)
-    expect(html).toEqual(expected)
   })
 });
